+++ CHANGELOG.md
@@ -5,7 +5,14 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.2 - Unreleased
+## Version 0.2.3 - Unreleased
+
+### Changed
+* Made bezier dependency optional
+* stats now returns video stats
+
+
+## Version 0.2.2 - Released 2021-05-10
 
 
 ### Added